♻️ [#104] Refactor: 파일 및 폴더 이름 변경
