+++ src/widgets/auth/components/index.ts
@@ -1,4 +1,4 @@
 export { default as SignUpField } from './sign-up-field/SignUpField'
 export { default as SignInField } from './sign-in-field/SignInField'
-export { default as SnsField } from './sns-field/SnsField'
+export { default as OauthField } from './oauth-field/OauthField'
 export { default as NicknameField } from './nickname-field/NicknameField'
